+++ packages/gov-ui-kit/src/core/utils/index.ts
@@ -2,3 +2,4 @@ export * from './clipboardUtils';
 export * from './formatterUtils';
 export * from './mergeRefs';
 export * from './responsiveUtils';
+export * from './ssrUtils';