+++ packages/gov-ui-kit/src/core/utils/ssrUtils/ssrUtils.ts
@@ -1,0 +1,8 @@
+class SsrUtils {
+    /**
+     * Checks if code is running on a server environment.
+     */
+    isServer = () => typeof window === 'undefined';
+}
+
+export const ssrUtils = new SsrUtils();