+++ packages/gov-ui-kit/src/modules/components/member/memberAvatar/memberAvatar.test.tsx
@@ -3,6 +3,7 @@ import * as blockies from 'blockies-ts';
 import * as viem from 'viem';
 import { normalize } from 'viem/ens';
 import * as wagmi from 'wagmi';
+import { ssrUtils } from '../../../../core';
 import { MemberAvatar, type IMemberAvatarProps } from './memberAvatar';
 
 describe('<MemberAvatar /> component', () => {
@@ -13,6 +14,7 @@ describe('<MemberAvatar /> component', () => {
     const useEnsAvatarMock = jest.spyOn(wagmi, 'useEnsAvatar');
     const getAddressMock = jest.spyOn(viem, 'getAddress');
     const blockiesCreateMock = jest.spyOn(blockies, 'create');
+    const isServerMock = jest.spyOn(ssrUtils, 'isServer');
 
     const createTestComponent = (props?: Partial<IMemberAvatarProps>) => {
         const completeProps: IMemberAvatarProps = { ...props };
@@ -22,12 +24,10 @@ describe('<MemberAvatar /> component', () => {
 
     beforeAll(() => {
         (window.Image as unknown) = class MockImage {
-            onload: () => void = () => {};
-            src: string = '';
+            onload = jest.fn();
+            src = '';
             constructor() {
-                setTimeout(() => {
-                    this.onload();
-                }, 100);
+                setTimeout(() => this.onload(), 100);
             }
         };
     });
@@ -50,39 +50,54 @@ describe('<MemberAvatar /> component', () => {
         useEnsAvatarMock.mockReset();
         getAddressMock.mockReset();
         blockiesCreateMock.mockReset();
+        isServerMock.mockReset();
     });
 
-    it('displays the avatar directly from avatarSrc prop and ensure no data fetching is attempted', async () => {
+    it('displays the avatar set as avatarSrc prop without resolving ens name, address or avatar', async () => {
         const avatarSrc = 'directAvatarUrl';
-
         render(createTestComponent({ avatarSrc }));
 
         await waitFor(() => expect(screen.getByRole('img')).toHaveAttribute('src', avatarSrc));
+        expect(useEnsAddressMock).toHaveBeenCalledWith(expect.objectContaining({ query: { enabled: false } }));
+        expect(useEnsNameMock).toHaveBeenCalledWith(expect.objectContaining({ query: { enabled: false } }));
+        expect(useEnsAvatarMock).toHaveBeenCalledWith(expect.objectContaining({ query: { enabled: false } }));
     });
 
     it('displays avatar for a valid ENS name and call related hooks with correct parameters', async () => {
         const ensName = 'vitalik.eth';
         const expectedAvatarUrl = 'ensAvatarUrl';
-
         useEnsAvatarMock.mockReturnValue({ data: expectedAvatarUrl, isLoading: false } as wagmi.UseEnsAvatarReturnType);
-
         render(createTestComponent({ ensName }));
 
         await waitFor(() => expect(screen.getByRole('img')).toHaveAttribute('src', expectedAvatarUrl));
-
+        expect(useEnsNameMock).toHaveBeenCalledWith(expect.objectContaining({ query: { enabled: false } }));
         expect(useEnsAvatarMock).toHaveBeenCalledWith(expect.objectContaining({ name: normalize(ensName) }));
     });
 
     it('displays avatar for a valid address and call related hooks with correct parameters', async () => {
         const address = '0x028F5Ca0b3A3A14e44AB8af660B53D1e428457e7';
         const expectedAvatarUrl = 'addressAvatarUrl';
-
         useEnsAvatarMock.mockReturnValue({ data: expectedAvatarUrl, isLoading: false } as wagmi.UseEnsAvatarReturnType);
-
         render(createTestComponent({ address }));
 
         await waitFor(() => expect(screen.getByRole('img')).toHaveAttribute('src', expectedAvatarUrl));
-
         expect(useEnsNameMock).toHaveBeenCalledWith(expect.objectContaining({ address }));
+    });
+
+    it('creates a blockies image and sets it as image fallback when having a valid address', () => {
+        const blockiesUrl = 'blockies-src-test';
+        blockiesCreateMock.mockReturnValue({ toDataURL: () => blockiesUrl } as HTMLCanvasElement);
+        const address = '0x028F5Ca0b3A3A14e44AB8af660B53D1e428457e7';
+        render(createTestComponent({ address }));
+
+        expect(blockiesCreateMock).toHaveBeenCalledWith(expect.objectContaining({ seed: address }));
+        expect(screen.getByRole('img')).toHaveAttribute('src', blockiesUrl);
+    });
+
+    it('does not create a blockies image on server environment', () => {
+        const address = '0x028F5Ca0b3A3A14e44AB8af660B53D1e428457e7';
+        isServerMock.mockReturnValue(true);
+        render(createTestComponent({ address }));
+        expect(blockiesCreateMock).not.toHaveBeenCalled();
     });
 });