+++ packages/gov-ui-kit/src/modules/components/member/memberAvatar/memberAvatar.tsx
@@ -3,7 +3,7 @@ import type React from 'react';
 import { type Address } from 'viem';
 import { normalize } from 'viem/ens';
 import { useEnsAddress, useEnsAvatar, useEnsName } from 'wagmi';
-import { Avatar, type IAvatarProps } from '../../../../core';
+import { Avatar, ssrUtils, type IAvatarProps } from '../../../../core';
 import { addressUtils, ensUtils } from '../../../utils';
 
 export interface IMemberAvatarProps extends Omit<IAvatarProps, 'fallback'> {
@@ -45,9 +45,10 @@ export const MemberAvatar: React.FC<IMemberAvatarProps> = (props) => {
     });
     const resolvedAvatarSrc = avatarSrc ?? ensAvatarData ?? undefined;
 
-    const blockiesSrc = resolvedAddress
-        ? blockies.create({ seed: addressUtils.getChecksum(resolvedAddress), scale: 8, size: 8 }).toDataURL()
-        : undefined;
+    const blockiesSrc =
+        resolvedAddress && !ssrUtils.isServer()
+            ? blockies.create({ seed: addressUtils.getChecksum(resolvedAddress), scale: 8, size: 8 }).toDataURL()
+            : undefined;
 
     const isLoading = avatarLoading || nameLoading || addressLoading;